properly format pip install path on Windows
it appears that on Windows based platforms the default pip install path(for Python < 2.7.9) is C:\Python27\Scripts however wagon assumed it is located along-side python inside os.path.dirname(sys.executable) which is C:\Python27 and therefore failed to work on Windows based platforms.

this patch resolves that issue.