rewrite platform/distro utils
the distro try/except importerror might be just this. Both distro's
and platform's `linux_distribution` have the same signature anyway.

Also, use distutils instead of the wheel.pep425tags call. This makes
us not necessarily depend on wheel itself (although actually doing
anything will still need wheel, of course. But wheel, similar to pip
and venv, doesn't consider its submodules public api, so this might
well change).
PEP 425 itself defines the platform as just that distutils call,
with those replaces.

For now, keep the distro requirement in setup.py as it was before.
In the future, maybe we should make it a hard requirement on py3.7+