Make linux_distribution optional (#135)
* Make linux_distribution optional

platform.linux_distribution doesn't exist on py3.8+, and
distro.linux_distribution doesn't exist if distro is not installed,
of course. This makes wagon break on import.

Instead, make the import optional, and only check that the function
is available in methods that actually use it, ie. create/repair
(not install)

* bump to 0.9.2