CY-3646 Rename manylinux wagons to linux_x86_64
Cloudify 5.1 (and earlier) doesn't handle manylinux wheels correctly.
It doesn't discover that manylinux wheels can be indeed installed
on linux platforms.

To allow wagons to still be installed on it, mark wagons that would
be manylinux, to be linux_x86_64 instead, which indeed are able
to be installed on Cloudify 5.1.

Also, allow the user to override the supported platform.

This is unfortunate but we'll have to deal with it for now :(