Update wagon.py (#142)
* Update wagon.py

The function _get_platform_for_set_of_wheels loops over a list of downloaded wheels to get the "real platform" of the wagon. This is prone to inconsistent results since the order of the wheel list varies from build to build. For now, we should just sort the wheels so that the results will be consistent for every list of wheels.

* Update setup.py

* Update CHANGES